requirements.txt
streamlit
numpy
pillow
opencv-python-headless
rembg
onnxruntime
